clippy and lint

+++ crates/llms/src/embeddings/mod.rs
@@ -40,5 +40,5 @@ pub trait Embed: Sync + Send {
     async fn embed(&mut self, input: EmbeddingInput) -> Result<Vec<Vec<f32>>>;
 
     /// Returns the size of the embedding vector returned by the model.
-    fn size(&self) -> usize;
+    fn size(&self) -> i32;
 }
